+++ src/test/scala/org/bitcoins/spvnode/serializers/messages/data/RawGetDataMessageSerializerTest.scala
@@ -1,6 +1,7 @@
 package org.bitcoins.spvnode.serializers.messages.data
 
 import org.bitcoins.core.crypto.DoubleSha256Digest
+import org.bitcoins.core.number.UInt64
 import org.bitcoins.core.protocol.CompactSizeUInt
 import org.bitcoins.spvnode.messages.MsgTx
 import org.bitcoins.spvnode.messages.data.Inventory
@@ -19,7 +20,7 @@ class RawGetDataMessageSerializerTest extends FlatSpec with MustMatchers {
 
   "RawGetDataMessageSerializer" must "read in a data message" in {
     val dataMsg = RawGetDataMessageSerializer.read(hex)
-    dataMsg.inventoryCount must be (CompactSizeUInt(2))
+    dataMsg.inventoryCount must be (CompactSizeUInt(UInt64(2)))
     dataMsg.inventories.head must be (Inventory(MsgTx, DoubleSha256Digest("de55ffd709ac1f5dc509a0925d0b1fc442ca034f224732e429081da1b621f55a")))
     dataMsg.inventories(1) must be (Inventory(MsgTx, DoubleSha256Digest("91d36d997037e08018262978766f24b8a055aaf1d872e94ae85e9817b2c68dc7")))
